requirements/requirements.txt
django==2.2
djangorestframework==3.9.2
django-environ==0.4.4
django-rest-swagger==2.2.0
psycopg2-binary==2.7.4
pillow==5.1.0
celery==4.2.0
django-celery-results==1.0.1

# dev packages
flake8==3.5.0
ipython==7.0.1
pytest==5.2.0
pytest-django==3.4.3
django-debug-toolbar==1.10.1
isort==4.3.4